pyproject.toml
[tool.poetry]
name = "event-management-agent"
version = "0.1.0"
description = "Simple agent which demonstrates how you can create conversational agents on top of some sinple REST interfaces."
authors = ["Gil Fernandes <gil.fernandes@onepointltd.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.3.7"
python-dotenv = "^1.0.0"
python-socketio = "^5.10.0"
aiohttp = "^3.9.1"
pydantic = "^2.5.2"
tomli = "^2.0.1"
tiktoken = "^0.5.2"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
